# backend/docs/openapi-schema.yml
openapi: 3.0.3
info:
  title: Ecommerce Web API
  version: 1.0.0
  description: REST API documentation for the Ecommerce Web platform, including authentication,
    catalog, cart, checkout, orders, payments, returns, inventory and administration.
  contact:
    name: Ecommerce Web Support
    email: support@example.com
paths:
...
  /api/payments/management/payments/{payment_number}/refund/:
    patch:
      operationId: payments_management_payments_refund_partial_update
      parameters:
      - in: path
        name: payment_number
        schema:
          type: string
        required: true
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: ''
  /api/payments/management/payments/{payment_number}/reject/:
    patch:
      operationId: payments_management_payments_reject_partial_update
      parameters:
      - in: path
        name: payment_number
        schema:
          type: string
        required: true
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: ''
  /api/payments/management/payments/{payment_number}/verify/:
    patch:
      operationId: payments_management_payments_verify_partial_update
      parameters:
      - in: path
        name: payment_number
        schema:
          type: string
        required: true
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: ''
...
  /api/shipments/management/shipments/{shipment_number}/update/:
    patch:
      operationId: shipments_management_shipments_update_partial_update
      parameters:
      - in: path
        name: shipment_number
        schema:
          type: string
        required: true
      tags:
      - shipments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
          description: ''
  /api/shipments/management/shipments/create/:
    post:
      operationId: shipments_management_shipments_create_create
      tags:
      - shipments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
          description: ''
...
components:
  schemas:
...
    MethodEnum:
      enum:
      - CASH_ON_DELIVERY
      - BANK_TRANSFER
      type: string
      description: |-
        * `CASH_ON_DELIVERY` - Cash on Delivery
        * `BANK_TRANSFER` - Bank Transfer
...
    Payment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        payment_number:
          type: string
          readOnly: true
        order_number:
          type: string
          readOnly: true
        customer_username:
          type: string
          readOnly: true
        method:
          allOf:
          - $ref: '#/components/schemas/MethodEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/PaymentStatusEnum'
          readOnly: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        transaction_reference:
          type: string
          readOnly: true
        proof:
          type: string
          format: uri
          nullable: true
          readOnly: true
        rejection_reason:
          type: string
          readOnly: true
        verified_by_username:
          type: string
          readOnly: true
          nullable: true
        submitted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        verified_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        rejected_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        refunded_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount
      - created_at
      - customer_username
      - id
      - method
      - order_number
      - payment_number
      - proof
      - refunded_at
      - rejected_at
      - rejection_reason
      - status
      - submitted_at
      - transaction_reference
      - updated_at
      - verified_at
      - verified_by_username
...
    PaymentStatusEnum:
      enum:
      - PENDING
      - PAID
      - FAILED
      - REFUNDED
      type: string
      description: |-
        * `PENDING` - Pending
        * `PAID` - Paid
        * `FAILED` - Failed
        * `REFUNDED` - Refunded
...
    Shipment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        shipment_number:
          type: string
          readOnly: true
        order_number:
          type: string
          readOnly: true
        customer_username:
          type: string
          readOnly: true
        customer_email:
          type: string
          format: email
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/Status408Enum'
          readOnly: true
        courier_name:
          type: string
          readOnly: true
        tracking_number:
          type: string
          readOnly: true
        estimated_delivery_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        shipped_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        delivered_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        events:
          type: array
          items:
            $ref: '#/components/schemas/ShipmentEvent'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - courier_name
      - created_at
      - customer_email
      - customer_username
      - delivered_at
      - estimated_delivery_date
      - events
      - id
      - order_number
      - shipment_number
      - shipped_at
      - status
      - tracking_number
      - updated_at
    ShipmentEvent:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/Status408Enum'
          readOnly: true
        message:
          type: string
          readOnly: true
        location:
          type: string
          readOnly: true
        created_by_username:
          type: string
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - created_by_username
      - id
      - location
      - message
      - status
...
    Status408Enum:
      enum:
      - READY
      - PICKED_UP
      - IN_TRANSIT
      - OUT_FOR_DELIVERY
      - DELIVERED
      - DELIVERY_FAILED
      - RETURNED
      - CANCELLED
      type: string
      description: |-
        * `READY` - Ready for Shipment
        * `PICKED_UP` - Picked Up
        * `IN_TRANSIT` - In Transit
        * `OUT_FOR_DELIVERY` - Out for Delivery
        * `DELIVERED` - Delivered
        * `DELIVERY_FAILED` - Delivery Failed
        * `RETURNED` - Returned
        * `CANCELLED` - Cancelled
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT